Sort npm package versions by their release date
This will help ensure releases aren't missed due to them being released
out of order.

It makes use of the "time" dictionary from the npm output, which has a
key of the release version and a value of when the version was
released.

The "versions" list only had a sorted list of all released versions,
which meant Prettier 2.8.0 for instance was skipped in
https://github.com/pre-commit/mirrors-prettier/issues/29, because
3.0.0-alpha.X releases already had been made before 2.8.0 was released.