Support user-provided hook names via `--name`
Until now, package_name and name were identical. This change allows the
user to provide a custom hook name which will then be used as the `name`
in the created hook.

As `name` is visible in a status message when
pre-commit runs, this allows to improve user experience by providing
messages like "Check XYZ formatting" instead of using the package name
like "xyz-format".

Usage of `name` and `package_name` in code are now disambiguated
through renaming.